refactor: extract bucket constants
