+++ src/components/SubscribeDialog.vue
@@ -60,7 +60,8 @@
 <script lang="ts">
 import { defineComponent, computed, ref, watch } from "vue";
 import { useDonationPresetsStore } from "stores/donationPresets";
-import { useBucketsStore, DEFAULT_BUCKET_ID } from "stores/buckets";
+import { useBucketsStore } from "stores/buckets";
+import { DEFAULT_BUCKET_ID } from "@/constants/buckets";
 import { useMintsStore } from "stores/mints";
 import { useUiStore } from "stores/ui";
 import {
