+++ src/constants/buckets.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_BUCKET_ID = 'unassigned'
+export const DEFAULT_BUCKET_NAME = 'Default'
