+++ src/pages/BucketDetail.vue
@@ -151,7 +151,8 @@
 <script setup lang="ts">
 import { computed, ref } from "vue";
 import { useRoute } from "vue-router";
-import { useBucketsStore, DEFAULT_BUCKET_ID } from "stores/buckets";
+import { useBucketsStore } from "stores/buckets";
+import { DEFAULT_BUCKET_ID } from "@/constants/buckets";
 import { useProofsStore } from "stores/proofs";
 import { useMintsStore } from "stores/mints";
 import { useUiStore } from "stores/ui";
