+++ src/stores/p2pk.ts
@@ -13,7 +13,8 @@ import { useProofsStore } from "./proofs";
 import { useMintsStore } from "./mints";
 import { useTokensStore } from "./tokens";
 import { useSendTokensStore } from "./sendTokensStore";
-import { DEFAULT_BUCKET_ID, useBucketsStore } from "./buckets";
+import { useBucketsStore } from "./buckets";
+import { DEFAULT_BUCKET_ID } from "@/constants/buckets";
 import { useLockedTokensStore } from "./lockedTokens";
 import { useSignerStore } from "./signer";
 import { notifyApiError, notifyError } from "src/js/notify";
