+++ src/stores/wallet.ts
@@ -14,7 +14,8 @@ import { useSendTokensStore } from "src/stores/sendTokensStore";
 import { usePRStore } from "./payment-request";
 import { useWorkersStore } from "./workers";
 import { useInvoicesWorkerStore } from "./invoicesWorker";
-import { DEFAULT_BUCKET_ID, useBucketsStore } from "./buckets";
+import { useBucketsStore } from "./buckets";
+import { DEFAULT_BUCKET_ID } from "@/constants/buckets";
 import { useLockedTokensStore } from "./lockedTokens";
 import { useInvoiceHistoryStore } from "./invoiceHistory";
 import { v4 as uuidv4 } from "uuid";
