+++ test/vitest/__tests__/bucketManagerArchive.spec.ts
@@ -26,9 +26,12 @@ vi.mock('../../../src/stores/buckets', () => ({
     editBucket: editBucketMock,
     deleteBucket: vi.fn(),
   }),
-  DEFAULT_BUCKET_ID: 'b1',
   COLOR_PALETTE: ['#fff'],
   hashColor: () => '#fff',
+}));
+
+vi.mock('../../../src/constants/buckets', () => ({
+  DEFAULT_BUCKET_ID: 'b1',
 }));
 
 vi.mock('../../../src/stores/mints', () => ({
