+++ test/wallet-store.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { useWalletStore } from 'stores/wallet';
-import { DEFAULT_BUCKET_ID } from 'stores/buckets';
+import { DEFAULT_BUCKET_ID } from '@/constants/buckets';
 
 let proofsStoreMock: any;
 let uiStoreMock: any;
